+++ gprof/symtab.h
@@ -41,11 +41,11 @@ typedef struct sym
       is_bb_head:1,		/* is this the head of a basic-blk? */
       mapped:1,			/* this symbol was mapped to another name */
       has_been_placed:1;	/* have we placed this symbol?  */
-    int ncalls;			/* how many times executed */
+    unsigned long ncalls;	/* how many times executed */
     int nuses;			/* how many times this symbol appears in
 				   a particular context */
     bfd_vma bb_addr[NBBS];	/* address of basic-block start */
-    int bb_calls[NBBS];		/* how many times basic-block was called */
+    unsigned long bb_calls[NBBS]; /* how many times basic-block was called */
     struct sym *next;		/* for building chains of syms */
     struct sym *prev;		/* for building chains of syms */
 
@@ -62,7 +62,7 @@ typedef struct sym
     /* call-graph specific info: */
     struct
       {
-	int self_calls;		/* how many calls to self */
+	unsigned long self_calls; /* how many calls to self */
 	double child_time;	/* cumulative ticks in children */
 	int index;		/* index in the graph list */
 	int top_order;		/* graph call chain top-sort order */
